refactor: Use 'args' property instead of '_args' attribute.
As described in Sympy's User Guide, one should never use internal
methods of variables, prefixed with '_' (as an example, it is explicitly
written that one should not use '_args', but use 'args' instead).

Here we apply this rule to the two main modules in the 'topology'
directory: 'basic' and 'domain'.