Allow turning off linearity checks in bilinear and linear forms (#153)
Add the boolean flag `check_linearity` to the constructors of
`BilinearForm` and `LinearForm`. The default value is `True`. If set to
`False`, the linearity checks are turned off completely. This is
different than the existing flag `ignore_linearity_errors`, which
controls whether we get a warning or an error when the linearity check
fails.

The new flag might be useful when the expression in the bilinear or
linear form is complicated, because then the linearity check calls the
SymPy `Expr` method `.expand(deep=True)`, which is recursive and can
take a very long time.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>